docshell/test/mochitest/test_bug404548.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=404548
-->
<head>
  <title>Test for Bug 404548</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=404548">Mozilla Bug 404548</a>
<p id="display">
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 404548 **/
var firstRemoved = false;
var secondHidden = false;

SimpleTest.waitForExplicitFinish();

var w = window.open("bug404548-subframe.html", "", "width=10,height=10");

function finishTest() {
  is(firstRemoved, true, "Should have removed iframe from the DOM");
  is(secondHidden, true, "Should have fired pagehide on second kid");
  w.close();
  SimpleTest.finish();
}
</script>
</pre>
</body>
</html>
